+++ lib/widgets/list_view.dart
@@ -684,7 +684,7 @@ class ListEntry extends StatelessWidget {
   Widget get listTile => Expanded(
       child: ListTile(
           contentPadding: contentPadding,
-          title: hero(title ?? MergeText(titleText, style: titleStyle)),
+          title: hero(title ?? BasisText(titleText, style: titleStyle)),
           subtitle: subtitle,
           leading: leading,
           trailing: child,